move FrameDataProvider into web package
