+++ src/main/java/de/focusshift/zeiterfassung/web/SignedInUserDto.java
@@ -1,4 +1,4 @@
-package de.focusshift.zeiterfassung.user;
+package de.focusshift.zeiterfassung.web;
 
 record SignedInUserDto(String fullName, String initials) {
 }
